+++ engine/src/game_types.h
@@ -8,8 +8,8 @@ struct game {
 
     // func pointers to game callbacks
     b8 (*initialize)(game* game_inst) = nullptr;
-    b8 (*update)(game* game_inst, f32 delta_time) = nullptr;
-    b8 (*render)(game* game_inst, f32 delta_time) = nullptr;
-    b8 (*on_resize)(game* game_inst, f32 delta_time) = nullptr;
+    b8 (*update)(game* game_inst, f64 delta_time) = nullptr;
+    b8 (*render)(game* game_inst, f64 delta_time) = nullptr;
+    b8 (*on_resize)(game* game_inst, f64 delta_time) = nullptr;
 };
